requirements.txt
pyserial
evdev
pyudev
opencv-python-headless
flask
flask-socketio
aiortс
 aiohttp